requirements.txt
appdirs==1.4.3
backcall==0.2.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipdb==0.13.4
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
parso==0.8.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
progress==1.5
prompt-toolkit==3.0.14
ptyprocess==0.7.0
Pygments==2.7.4
pyparsing==2.4.6
python-decouple==3.4
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
traitlets==5.0.5
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
